Detach Filter calls into FiltersMixin, detach FilterType and getter.
